Move supplier questions views into separate module
